Projects/task_manager_with_mongodb/pyproject.toml
[project]
name = "Task Manager with MongoDB"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai-agents>=0.2.9",
    "pymongo[srv]>=4.14.1",
    "python-dotenv>=1.1.1",
]
